pyproject.toml
[project]
name = "lct-dendrology"
version = "0.1.0"
description = "Telegram bot that accepts images and replies with text"
authors = [
    {name = "alex"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "python-telegram-bot[rate-limiter] (>=22.4,<23.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi (>=0.104.0,<1.0.0)",
    "uvicorn[standard] (>=0.24.0,<1.0.0)",
    "python-multipart (>=0.0.6,<1.0.0)",
    "aiohttp (>=3.9.0,<4.0.0)",
    "pydantic-settings (>=2.0.0,<3.0.0)"
]

[project.optional-dependencies]
test = [
    "pytest (>=7.0.0,<8.0.0)",
    "pytest-asyncio (>=0.21.0,<1.0.0)",
    "httpx (>=0.24.0,<1.0.0)",
    "pillow (>=10.0.0,<11.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
